pyproject.toml
[project]
name = "grim"
version = "0.1.0"
description = "GRIM — General Recursive Intelligence Machine"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langgraph-checkpoint-sqlite>=3.0",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "langchain-mcp-adapters>=0.1",
    "mcp>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
grim = "core.__main__:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["core*", "server*"]
